Fix circular dependency between writable and duplex
