+++ node/internal/streams/_utils.ts
@@ -1,0 +1,9 @@
+import { Buffer } from "../../buffer.ts";
+
+export function _uint8ArrayToBuffer(chunk: Uint8Array) {
+  return Buffer.from(
+    chunk.buffer,
+    chunk.byteOffset,
+    chunk.byteLength,
+  );
+}
